[ty] Refine gradual generic intersection handling
