+++ crates/ty_python_semantic/src/types/subclass_of.rs
@@ -6,7 +6,7 @@ use crate::types::relation::{DisjointnessChecker, TypeRelationChecker};
 use crate::types::variance::VarianceInferable;
 use crate::types::{
     ApplyTypeMappingVisitor, BoundTypeVarInstance, ClassLiteral, ClassType, DynamicType,
-    FindLegacyTypeVarsVisitor, IntersectionType, KnownClass, MaterializationKind,
+    FindLegacyTypeVarsVisitor, IntersectionBuilder, KnownClass, MaterializationKind,
     MemberLookupPolicy, SpecialFormType, Type, TypeContext, TypeMapping, TypeVarBoundOrConstraints,
     TypeVarVariance, TypedDictType, UnionType, todo_type,
 };
@@ -90,13 +90,12 @@ impl<'db> SubclassOfType<'db> {
                     .map(|element| Self::try_from_instance(db, *element)),
             ),
             Type::Intersection(intersection) if intersection.iter_negative(db).next().is_none() => {
-                Some(IntersectionType::from_elements(
-                    db,
-                    intersection
-                        .iter_positive(db)
-                        .map(|element| Self::try_from_instance(db, element))
-                        .collect::<Option<Vec<_>>>()?,
-                ))
+                intersection
+                    .iter_positive(db)
+                    .try_fold(IntersectionBuilder::new(db), |builder, element| {
+                        Some(builder.add_positive(Self::try_from_instance(db, element)?))
+                    })
+                    .map(IntersectionBuilder::build)
             }
             Type::ProtocolInstance(protocol) => Some(protocol.to_meta_type(db)),
             _ => SubclassOfInner::try_from_instance(db, ty)
